refactor: split frontend large files + fix any types
- #11: Split Graph.tsx (1621→92), Graph3D.tsx (1236→616), Traces.tsx (975→345)
  Extract Canvas2D, Canvas3D, TraceDetailView components
  Extract graphMath, graphClustering, graphTypes utilities
- #14: Replace all any types with proper interfaces
  - oracle.ts: GraphNodeData/GraphLinkData interfaces
  - useHandTracking.ts: unknown cast instead of any
  - Events.tsx: unknown error with instanceof check
